requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo[srv]==4.5.0
python-dotenv==1.0.0
motor==3.2.0
python-decouple==3.8
python-multipart==0.0.6